Split sidebar hover reconciler tests
